Moved __mocks__ to root folder
