+++ __mocks__/react-i18next.ts
@@ -1,0 +1,13 @@
+import noop from 'lodash/noop'
+
+export const useTranslation = () => {
+  return {
+    t: (str: string) => str,
+    i18n: {
+      changeLanguage: () => new Promise(noop),
+      language: 'it',
+    },
+  }
+}
+
+export const Trans = ({ children }) => children
