+++ setupTests.js
@@ -1,7 +1,6 @@
 import { expect, afterEach, vi } from 'vitest'
 import { cleanup } from '@testing-library/react'
 import matchers from '@testing-library/jest-dom/matchers'
-import noop from 'lodash/noop'
 
 // extends Vitest's expect method with methods from react-testing-library
 expect.extend(matchers)
@@ -12,17 +11,5 @@ afterEach(() => {
 })
 
 vi.stubGlobal('scroll', vi.fn())
-
-vi.mock('react-i18next', () => ({
-  // this mock makes sure any components using the translate hook can use it without a warning being shown
-  useTranslation: () => {
-    return {
-      t: (str) => str,
-      i18n: {
-        changeLanguage: () => new Promise(noop),
-        language: 'it',
-      },
-    }
-  },
-  Trans: ({ children }) => children,
-}))
+vi.mock('zustand')
+vi.mock('react-i18next')
