+++ src/api/react-query-wrappers/__tests__/useQueryWrapper.test.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { renderHook, screen } from '@testing-library/react'
 import { useQueryWrapper } from '../useQueryWrapper'
 import { QueryClientProvider } from '@tanstack/react-query'
-import { queryClientMock } from '@/__mocks__/query-client.mock'
+import { queryClientMock } from '@/utils/testing.utils'
 import { vi } from 'vitest'
 import * as hooks from '@/hooks/useJwt'
 import { act } from 'react-dom/test-utils'
